refactor: move mcpswap package under internal/x/
